[Clean] Migrate Bookmarks properties to WritableIntDefPropertyKey
Migrate BookmarkManagerProperties.LOCATION,
BookmarkToolbarProperties.BOOKMARK_UI_MODE, and
ImprovedBookmarkRowProperties.END_IMAGE_VISIBILITY to
WritableIntDefPropertyKey.

TAG=agy
CONV=de2e9bd7-30e4-4425-bcaa-6646139733fa

Bug: 547916752
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8266700
Commit-Queue: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1687254}

+++ chrome/browser/bookmarks/android/junit/src/org/chromium/chrome/browser/bookmarks/BookmarkToolbarMediatorTest.java
@@ -210,18 +210,14 @@ public class BookmarkToolbarMediatorTest {
     public void onStateChangedUpdatesModel() {
         mMediator.onUiModeChanged(BookmarkUiMode.LOADING);
         assertEquals(
-                BookmarkUiMode.LOADING,
-                mModel.get(BookmarkToolbarProperties.BOOKMARK_UI_MODE).intValue());
+                BookmarkUiMode.LOADING, mModel.get(BookmarkToolbarProperties.BOOKMARK_UI_MODE));
 
         mMediator.onUiModeChanged(BookmarkUiMode.SEARCHING);
         assertEquals(
-                BookmarkUiMode.SEARCHING,
-                mModel.get(BookmarkToolbarProperties.BOOKMARK_UI_MODE).intValue());
+                BookmarkUiMode.SEARCHING, mModel.get(BookmarkToolbarProperties.BOOKMARK_UI_MODE));
 
         mMediator.onUiModeChanged(BookmarkUiMode.FOLDER);
-        assertEquals(
-                BookmarkUiMode.FOLDER,
-                mModel.get(BookmarkToolbarProperties.BOOKMARK_UI_MODE).intValue());
+        assertEquals(BookmarkUiMode.FOLDER, mModel.get(BookmarkToolbarProperties.BOOKMARK_UI_MODE));
     }
 
     @Test
